requirements.txt
# SQLite3 входит в стандартную библиотеку Python

# ============================================
# Backend API Framework
# ============================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# ============================================
# Database ORM
# ============================================
sqlalchemy>=2.0.0

# ============================================
# Data Validation
# ============================================
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ============================================
# Authentication & Security
# ============================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# ============================================
# File Uploads
# ============================================
python-multipart>=0.0.6

# ============================================
# Telegram Bot Framework
# ============================================
aiogram>=3.2.0
aiohttp>=3.9.0

# ============================================
# Background Task Scheduler
# ============================================
apscheduler>=3.10.0

# ============================================
# Environment Variables
# ============================================
python-dotenv>=1.0.0

# ============================================
# Database Migrations (Optional)
# ============================================
alembic>=1.12.0

# ============================================
# Для компиляции в .exe файлы
# ============================================
pyinstaller>=5.0